Revise isolation tag logic (#125)
* adjust to fit current production style usage ("dev" tag when isolating)
* improve tag value displays

Note that this change breaks backward compatibility of isolation on some dev environments
that do not support the "dev" tag. If you see error indicating that the "dev" tag is not supported,
the workaround is to package the solution into a zip file with the desired isolation (non-stable) tag and then
push or validate with the "stable" flag:
1. `fsoc solution package --solution-bundle mysolution.zip --tag joe`
1. `fsoc solution push --solution-bundle mysolution.zip --tag stable`

The operation of push for "stable" and for non-isolated solutions is not affected.